+++ IPv4/Converter.php
@@ -25,6 +25,9 @@ use function ltrim;
 use function preg_match;
 use function str_ends_with;
 use function substr;
+use const FILTER_FLAG_IPV4;
+use const FILTER_FLAG_IPV6;
+use const FILTER_VALIDATE_IP;
 
 final class Converter
 {
@@ -42,6 +45,9 @@ final class Converter
         '/^0(?<number>[0-7]*)$/' => 8,
         '/^(?<number>\d+)$/' => 10,
     ];
+
+    private const IPV6_6TO4_PREFIX = '2002:';
+    private const IPV4_MAPPED_PREFIX = '::ffff:';
 
     private readonly mixed $maxIPv4Number;
 
@@ -95,7 +101,58 @@ final class Converter
 
     public function isIpv4(Stringable|string|null $host): bool
     {
-        return null !== $this->toDecimal($host);
+        if (null === $host) {
+            return false;
+        }
+
+        if (null !== $this->toDecimal($host)) {
+            return true;
+        }
+
+        $host = (string) $host;
+        if (false === filter_var($host, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
+            return false;
+        }
+
+        $ipAddress = strtolower((string) inet_ntop((string) inet_pton($host)));
+        if (str_starts_with($ipAddress, self::IPV4_MAPPED_PREFIX)) {
+            return false !== filter_var(substr($ipAddress, 7), FILTER_VALIDATE_IP, FILTER_FLAG_IPV4);
+        }
+
+        if (!str_starts_with($ipAddress, self::IPV6_6TO4_PREFIX)) {
+            return false;
+        }
+
+        $hexParts = explode(':', substr($ipAddress, 5, 9));
+
+        return count($hexParts) > 1
+            && false !== long2ip((int) hexdec($hexParts[0]) * 65536 + (int) hexdec($hexParts[1]));
+    }
+
+    public function to6to4(Stringable|string|null $host): ?string
+    {
+        $host = $this->toDecimal($host);
+        if (null === $host) {
+            return null;
+        }
+
+        /** @var array<string> $parts */
+        $parts = array_map(
+            fn (string $part): string => sprintf('%02x', $part),
+            explode('.', $host)
+        );
+
+        return '['.self::IPV6_6TO4_PREFIX . $parts[0] . $parts[1] . ':' . $parts[2] . $parts[3] . '::]';
+    }
+
+    public function toIPv4MappedIPv6(Stringable|string|null $host): ?string
+    {
+        $host = $this->toDecimal($host);
+
+        return match ($host) {
+            null => null,
+            default => '['.self::IPV4_MAPPED_PREFIX.$host.']',
+        };
     }
 
     public function toOctal(Stringable|string|null $host): ?string
@@ -133,6 +190,29 @@ final class Converter
     public function toDecimal(Stringable|string|null $host): ?string
     {
         $host = (string) $host;
+        if (str_starts_with($host, '[') && str_ends_with($host, ']')) {
+            $host = substr($host, 1, -1);
+            if (false === filter_var($host, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
+                return null;
+            }
+
+            $ipAddress = strtolower((string) inet_ntop((string) inet_pton($host)));
+            if (str_starts_with($ipAddress, self::IPV4_MAPPED_PREFIX)) {
+                return substr($ipAddress, 7);
+            }
+
+            if (!str_starts_with($ipAddress, self::IPV6_6TO4_PREFIX)) {
+                return null;
+            }
+
+            $hexParts = explode(':', substr($ipAddress, 5, 9));
+
+            return (string) match (true) {
+                count($hexParts) < 2 => null,
+                default => long2ip((int) hexdec($hexParts[0]) * 65536 + (int) hexdec($hexParts[1])),
+            };
+        }
+
         if (1 !== preg_match(self::REGEXP_IPV4_HOST, $host)) {
             return null;
         }