+++ IPv6/Converter.php
@@ -6,6 +6,7 @@ namespace League\Uri\IPv6;
 
 use Stringable;
 use ValueError;
+
 use const FILTER_FLAG_IPV6;
 use const FILTER_VALIDATE_IP;
 
@@ -18,22 +19,28 @@ use function unpack;
 
 final class Converter
 {
-    public static function compressIp(string $ipv6): string
-    {
-        if (false === filter_var($ipv6, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
-            throw new ValueError('The submitted IP is not a valid IPv6 address.');
-        }
+    /**
+     * Significant 10 bits of IP to detect Zone ID regular expression pattern.
+     *
+     * @var string
+     */
+    private const HOST_ADDRESS_BLOCK = "\xfe\x80";
 
-        return (string) inet_ntop((string) inet_pton($ipv6));
+    public static function compressIp(string $ipAddress): string
+    {
+        return match (filter_var($ipAddress, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
+            false => throw new ValueError('The submitted IP is not a valid IPv6 address.'),
+            default =>  strtolower((string) inet_ntop((string) inet_pton($ipAddress))),
+        };
     }
 
-    public static function expandIp(string $ipv6): string
+    public static function expandIp(string $ipAddress): string
     {
-        if (false === filter_var($ipv6, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
+        if (false === filter_var($ipAddress, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
             throw new ValueError('The submitted IP is not a valid IPv6 address.');
         }
 
-        $hex = (array) unpack("H*hex", (string) inet_pton($ipv6));
+        $hex = (array) unpack("H*hex", (string) inet_pton($ipAddress));
 
         return implode(':', str_split(strtolower($hex['hex'] ?? ''), 4));
     }
@@ -41,14 +48,14 @@ final class Converter
     public static function compress(Stringable|string|null $host): ?string
     {
         $components = self::parse($host);
-        if (null === $components['ipv6']) {
+        if (null === $components['ipAddress']) {
             return match ($host) {
                 null => $host,
                 default => (string) $host,
             };
         }
 
-        $components['ipv6'] = self::compressIp($components['ipv6']);
+        $components['ipAddress'] = self::compressIp($components['ipAddress']);
 
         return self::build($components);
     }
@@ -56,24 +63,28 @@ final class Converter
     public static function expand(Stringable|string|null $host): ?string
     {
         $components = self::parse($host);
-        if (null === $components['ipv6']) {
+        if (null === $components['ipAddress']) {
             return match ($host) {
                 null => $host,
                 default => (string) $host,
             };
         }
 
-        $components['ipv6'] = self::expandIp($components['ipv6']);
+        $components['ipAddress'] = self::expandIp($components['ipAddress']);
 
         return self::build($components);
     }
 
     private static function build(array $components): string
     {
-        $components['ipv6'] ??= null;
+        $components['ipAddress'] ??= null;
         $components['zoneIdentifier'] ??= null;
 
-        return '['.$components['ipv6'].match ($components['zoneIdentifier']) {
+        if (null === $components['ipAddress']){
+            return '';
+        }
+
+        return '['.$components['ipAddress'].match ($components['zoneIdentifier']) {
             null => '',
             default => '%'.$components['zoneIdentifier'],
         }.']';
@@ -82,32 +93,36 @@ final class Converter
     /**]
      * @param Stringable|string|null $host
      *
-     * @return array{ipv6:?string, zoneIdentifier:?string}
+     * @return array{ipAddress:string|null, zoneIdentifier:string|null}
      */
     private static function parse(Stringable|string|null $host): array
     {
         if ($host === null) {
-            return ['ipv6' => null, 'zoneIdentifier' => null];
+            return ['ipAddress' => null, 'zoneIdentifier' => null];
         }
 
         $host = (string) $host;
         if ($host === '') {
-            return ['ipv6' => null, 'zoneIdentifier' => null];
+            return ['ipAddress' => null, 'zoneIdentifier' => null];
         }
 
         if (!str_starts_with($host, '[')) {
-            return ['ipv6' => null, 'zoneIdentifier' => null];
+            return ['ipAddress' => null, 'zoneIdentifier' => null];
         }
 
         if (!str_ends_with($host, ']')) {
-            return ['ipv6' => null, 'zoneIdentifier' => null];
+            return ['ipAddress' => null, 'zoneIdentifier' => null];
         }
 
         [$ipv6, $zoneIdentifier] = explode('%', substr($host, 1, -1), 2) + [1 => null];
         if (false === filter_var($ipv6, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
-            return ['ipv6' => null, 'zoneIdentifier' => null];
+            return ['ipAddress' => null, 'zoneIdentifier' => null];
         }
 
-        return ['ipv6' => $ipv6, 'zoneIdentifier' => $zoneIdentifier];
+        return match (true) {
+            null === $zoneIdentifier,
+            is_string($ipv6) && str_starts_with((string)inet_pton($ipv6), self::HOST_ADDRESS_BLOCK) =>  ['ipAddress' => $ipv6, 'zoneIdentifier' => $zoneIdentifier],
+            default => ['ipAddress' => null, 'zoneIdentifier' => null],
+        };
     }
 }